requirements.txt
asgiref==3.8.1
attrs==25.1.0
azure-core==1.32.0
azure-storage-blob==12.24.1
black==25.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
cryptography==44.0.0
diff-match-patch==20241021
Django==5.1.5
django-admin-rangefilter==0.13.2
django-filter==24.3
django-import-export==4.3.4
django-jazzmin==3.0.1
django-session-timeout==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==35.2.0
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isodate==0.7.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.0
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
